\name{get_rand_int}
\alias{get_rand_int}
\title{Generate a random integer between two integers}
\usage{
  get_rand_int(min = 0, max = 1)
}
\arguments{
  \item{min}{minimum value}

  \item{max}{maximum value}
}
\description{
  Generate a random integer between two integers
}
\examples{
get_rand_int(min=0, max=10)
}

